requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==4.0.0
torch==2.0.0
transformers==4.30.0
pillow==9.5.0
numpy==1.24.3
python-dotenv==0.19.0
requests==2.28.1
opencv-python==4.7.0.72
scikit-learn==1.3.0
pyttsx3==2.90
sounddevice==0.4.5
